+++ Projects/Domain/UseCase/Implementations/ChallengeUseCaseImpl.swift
@@ -15,13 +15,19 @@ import UseCase
 import Repository
 
 public struct ChallengeUseCaseImpl: ChallengeUseCase {
-  private let repository: ChallengeRepository
+  private let challengeRepository: ChallengeRepository
+  private let feedRepository: FeedRepository
   private let authRepository: AuthRepository
   private let challengeProveMemberCountRelay = BehaviorRelay<Int>(value: 0)
   public let challengeProveMemberCount: Infallible<Int>
   
-  public init(repository: ChallengeRepository, authRepository: AuthRepository) {
-    self.repository = repository
+  public init(
+    challengeRepository: ChallengeRepository,
+    feedRepository: FeedRepository,
+    authRepository: AuthRepository
+  ) {
+    self.challengeRepository = challengeRepository
+    self.feedRepository = feedRepository
     self.authRepository = authRepository
     
     self.challengeProveMemberCount = challengeProveMemberCountRelay.asInfallible()
@@ -31,7 +37,7 @@ public struct ChallengeUseCaseImpl: ChallengeUseCase {
 // MARK: - Fetch Methods
 public extension ChallengeUseCaseImpl {
   func fetchChallengeDetail(id: Int) -> Single<ChallengeDetail> {
-    return repository.fetchChallengeDetail(id: id)
+    return challengeRepository.fetchChallengeDetail(id: id)
   }
   
   func isLogIn() async -> Bool {
@@ -39,7 +45,7 @@ public extension ChallengeUseCaseImpl {
   }
   
   func fetchFeeds(id: Int, page: Int, size: Int, orderType: ChallengeFeedsOrderType) async throws -> PageFeeds {
-    let result = try await repository.fetchFeeds(id: id, page: page, size: size, orderType: orderType)
+    let result = try await feedRepository.fetchFeeds(id: id, page: page, size: size, orderType: orderType)
     
     if challengeProveMemberCountRelay.value != result.memberCount {
       challengeProveMemberCountRelay.accept(result.memberCount)
@@ -49,22 +55,22 @@ public extension ChallengeUseCaseImpl {
   }
   
   func fetchChallengeDescription(id: Int) -> Single<ChallengeDescription> {
-    return repository.fetchChallengeDescription(challengeId: id)
+    return challengeRepository.fetchChallengeDescription(challengeId: id)
   }
   
   func fetchChallengeMembers(challengeId: Int) -> Single<[ChallengeMember]> {
-    return repository.fetchChallengeMembers(challengeId: challengeId)
+    return challengeRepository.fetchChallengeMembers(challengeId: challengeId)
   }
 }
 
 // MARK: - Upload & Update Methods
 public extension ChallengeUseCaseImpl {
   func joinPrivateChallnege(id: Int, code: String) async throws {
-    try await repository.joinPrivateChallnege(id: id, code: code).value
+    try await challengeRepository.joinPrivateChallnege(id: id, code: code).value
   }
   
   func joinPublicChallenge(id: Int) -> Single<Void> {
-    return repository.joinPublicChallenge(id: id)
+    return challengeRepository.joinPublicChallenge(id: id)
   }
   
   func uploadChallengeFeedProof(id: Int, image: UIImageWrapper) async throws {
@@ -72,7 +78,7 @@ public extension ChallengeUseCaseImpl {
       throw APIError.challengeFailed(reason: .fileTooLarge)
     }
     
-    try await repository.uploadChallengeFeedProof(
+    try await challengeRepository.uploadChallengeFeedProof(
       id: id,
       image: data,
       imageType: type
@@ -80,15 +86,15 @@ public extension ChallengeUseCaseImpl {
   }
   
   func updateLikeState(challengeId: Int, feedId: Int, isLike: Bool) async throws {
-    return try await repository.updateLikeState(challengeId: challengeId, feedId: feedId, isLike: isLike)
+    return try await feedRepository.updateLikeState(challengeId: challengeId, feedId: feedId, isLike: isLike)
   }
   
   func isProve(challengeId: Int) async throws -> Bool {
-    return try await repository.isProve(challengeId: challengeId)
+    return try await challengeRepository.isProve(challengeId: challengeId)
   }
   
   func updateChallengeGoal(_ goal: String, challengeId: Int) -> Single<Void> {
-    return repository.updateChallengeGoal(goal, challengeId: challengeId)
+    return challengeRepository.updateChallengeGoal(goal, challengeId: challengeId)
   }
 }
